follow graph 통합 테스트를 분리한다
