+++ apps/api/tests/integration/graphql/profile.test.ts
@@ -16,7 +16,7 @@ import {
 import { postContentDocumentFromText } from '@kosmo/core/post-content/server';
 import { isConfiguredLocalProfile } from '@kosmo/core/profile';
 import { normalizeHandle } from '@kosmo/core/utils';
-import { and, asc, count, eq, ne } from 'drizzle-orm';
+import { and, count, eq, ne } from 'drizzle-orm';
 import { Hono } from 'hono';
 import type * as CoreDb from '@kosmo/core/db';
 import type * as CoreSeed from '@kosmo/core/db/seed';
@@ -212,276 +212,6 @@ describe('GraphQL remote profile boundary', () => {
 
     assertNoGraphQLErrors(result);
     assert.equal(result.data?.node?.instance.kind, 'LOCAL');
-  });
-
-  test('reads follow graph data for active profiles on another local instance', async () => {
-    const auth = await createAuthenticatedSession();
-    const otherLocalInstance = await createLocalInstance({ domain: 'other-local.example' });
-    const otherLocal = await createProfile({
-      handle: 'other-local',
-      instanceId: otherLocalInstance.id,
-    });
-
-    await db.insert(ProfileFollows).values([
-      { followerProfileId: auth.profile.id, followeeProfileId: otherLocal.id },
-      { followerProfileId: otherLocal.id, followeeProfileId: auth.profile.id },
-    ]);
-
-    const result = await requestGraphQL<{
-      node: {
-        followers: { edges: Array<{ node: { id: string } }> };
-        following: { edges: Array<{ node: { id: string } }> };
-        viewerFollow: { id: string } | null;
-        viewerState: { follow: { id: string } | null; isSelf: boolean } | null;
-      } | null;
-    }>(
-      `query OtherLocalFollowGraph($id: ID!) {
-        node(id: $id) {
-          ... on Profile {
-            followers(first: 10) { edges { node { id } } }
-            following(first: 10) { edges { node { id } } }
-            viewerFollow { id }
-            viewerState { isSelf follow { id } }
-          }
-        }
-      }`,
-      { id: otherLocal.id },
-      auth.token,
-    );
-
-    assertNoGraphQLErrors(result);
-    assert.equal(result.data?.node?.followers.edges.length, 1);
-    assert.equal(result.data?.node?.following.edges.length, 1);
-    assert.equal(
-      result.data?.node?.viewerFollow?.id,
-      result.data?.node?.followers.edges[0]?.node.id,
-    );
-    assert.deepEqual(result.data?.node?.viewerState, {
-      follow: result.data?.node?.viewerFollow,
-      isSelf: false,
-    });
-  });
-
-  test('does not expose pending requests as local follow relationships', async () => {
-    const auth = await createAuthenticatedSession();
-    const otherLocalInstance = await createLocalInstance({ domain: 'pending-local.example' });
-    const otherLocal = await createProfile({
-      handle: 'pending-local',
-      instanceId: otherLocalInstance.id,
-    });
-    await db.insert(ProfileFollowRequests).values([
-      { followerProfileId: auth.profile.id, followeeProfileId: otherLocal.id },
-      { followerProfileId: otherLocal.id, followeeProfileId: auth.profile.id },
-    ]);
-
-    const result = await requestGraphQL<{
-      node: {
-        followers: { edges: Array<{ node: { id: string } }> };
-        following: { edges: Array<{ node: { id: string } }> };
-        viewerFollow: { id: string } | null;
-        viewerState: { follow: { id: string } | null; isSelf: boolean } | null;
-      } | null;
-    }>(
-      `query PendingLocalFollowGraph($id: ID!) {
-        node(id: $id) {
-          ... on Profile {
-            followers(first: 10) { edges { node { id } } }
-            following(first: 10) { edges { node { id } } }
-            viewerFollow { id }
-            viewerState { isSelf follow { id } }
-          }
-        }
-      }`,
-      { id: otherLocal.id },
-      auth.token,
-    );
-
-    assertNoGraphQLErrors(result);
-    assert.deepEqual(result.data?.node, {
-      followers: { edges: [] },
-      following: { edges: [] },
-      viewerFollow: null,
-      viewerState: { follow: null, isSelf: false },
-    });
-  });
-
-  test('reads visible established relationships and stored counts for a remote profile', async () => {
-    const remoteInstance = await createRemoteInstance();
-    const remote = await createProfile({
-      followersCount: 41,
-      followingCount: 43,
-      handle: 'remote-follow-graph',
-      instanceId: remoteInstance.id,
-    });
-    const publicFollower = await createProfile({
-      handle: 'public-follower',
-      instanceId: localInstanceId,
-    });
-    const publicFollowee = await createProfile({
-      handle: 'public-followee',
-      instanceId: localInstanceId,
-    });
-    const [followerFollow, followingFollow] = await db
-      .insert(ProfileFollows)
-      .values([
-        { followerProfileId: publicFollower.id, followeeProfileId: remote.id },
-        { followerProfileId: remote.id, followeeProfileId: publicFollowee.id },
-      ])
-      .returning();
-    const result = await requestFollowGraph(`remote-follow-graph@${remoteDomain}`);
-
-    assertNoGraphQLErrors(result);
-    assert.deepEqual(result.data?.profileByHandle, {
-      followers: { edges: [{ node: { id: followerFollow!.id } }] },
-      followersCount: 41,
-      following: { edges: [{ node: { id: followingFollow!.id } }] },
-      followingCount: 43,
-      viewerFollow: null,
-      viewerState: null,
-    });
-  });
-
-  test("shows the viewer's own approval-required relationships", async () => {
-    const auth = await createAuthenticatedSession();
-    await db
-      .update(Profiles)
-      .set({ followPolicy: ProfileFollowPolicy.APPROVAL_REQUIRED })
-      .where(eq(Profiles.id, auth.profile.id));
-    const remoteInstance = await createRemoteInstance();
-    const remote = await createProfile({
-      handle: 'viewer-private-follow-graph',
-      instanceId: remoteInstance.id,
-    });
-    const [viewerFollow, reverseViewerFollow] = await db
-      .insert(ProfileFollows)
-      .values([
-        { followerProfileId: auth.profile.id, followeeProfileId: remote.id },
-        { followerProfileId: remote.id, followeeProfileId: auth.profile.id },
-      ])
-      .returning();
-
-    const result = await requestFollowGraph(
-      `viewer-private-follow-graph@${remoteDomain}`,
-      auth.token,
-    );
-
-    assertNoGraphQLErrors(result);
-    assert.deepEqual(result.data?.profileByHandle, {
-      followers: { edges: [{ node: { id: viewerFollow!.id } }] },
-      followersCount: 0,
-      following: { edges: [{ node: { id: reverseViewerFollow!.id } }] },
-      followingCount: 0,
-      viewerFollow: { id: viewerFollow!.id },
-      viewerState: { follow: { id: viewerFollow!.id }, isSelf: false },
-    });
-  });
-
-  test('excludes approval-required relationships from a remote follow graph', async () => {
-    await assertRemoteFollowGraphHidesCounterparts({
-      followPolicy: ProfileFollowPolicy.APPROVAL_REQUIRED,
-    });
-  });
-
-  test('excludes inactive profile relationships from a remote follow graph', async () => {
-    await assertRemoteFollowGraphHidesCounterparts({ profileState: ProfileState.DISABLED });
-  });
-
-  test('excludes suspended instance relationships from a remote follow graph', async () => {
-    await assertRemoteFollowGraphHidesCounterparts({
-      instanceState: InstanceState.SUSPENDED,
-    });
-  });
-
-  test('returns viewer follow for an outgoing established relationship', async () => {
-    const { followId, result } = await requestViewerFollowScenario('outgoing');
-
-    assertNoGraphQLErrors(result);
-    assert.deepEqual(result.data?.profileByHandle?.viewerFollow, { id: followId });
-    assert.deepEqual(result.data?.profileByHandle?.viewerState, {
-      follow: { id: followId },
-      isSelf: false,
-    });
-  });
-
-  test('does not return viewer follow for a reverse-only relationship', async () => {
-    const { result } = await requestViewerFollowScenario('reverse');
-
-    assertNoGraphQLErrors(result);
-    assert.equal(result.data?.profileByHandle?.viewerFollow, null);
-    assert.deepEqual(result.data?.profileByHandle?.viewerState, {
-      follow: null,
-      isSelf: false,
-    });
-  });
-
-  test('does not expose pending requests as follow relationships', async () => {
-    const { result } = await requestViewerFollowScenario('pending', {
-      followersCount: 47,
-      followingCount: 53,
-    });
-
-    assertNoGraphQLErrors(result);
-    assert.deepEqual(result.data?.profileByHandle, {
-      followers: { edges: [] },
-      followersCount: 47,
-      following: { edges: [] },
-      followingCount: 53,
-      viewerFollow: null,
-      viewerState: { follow: null, isSelf: false },
-    });
-  });
-
-  test('reads a remote follow graph without network or database writes', async () => {
-    const auth = await createAuthenticatedSession();
-    const remoteInstance = await createRemoteInstance();
-    const remote = await createProfile({
-      followersCount: 61,
-      followingCount: 67,
-      handle: 'read-only-remote-follow-graph',
-      instanceId: remoteInstance.id,
-    });
-    const pendingFollower = await createProfile({
-      handle: 'pending-follower',
-      instanceId: localInstanceId,
-    });
-    await db
-      .insert(ProfileFollows)
-      .values({ followerProfileId: auth.profile.id, followeeProfileId: remote.id });
-    await db
-      .insert(ProfileFollowRequests)
-      .values({ followerProfileId: pendingFollower.id, followeeProfileId: remote.id });
-
-    const readFollowGraphState = async () =>
-      JSON.stringify({
-        instances: await db.select().from(Instances).orderBy(asc(Instances.id)),
-        profiles: await db.select().from(Profiles).orderBy(asc(Profiles.id)),
-        follows: await db.select().from(ProfileFollows).orderBy(asc(ProfileFollows.id)),
-        requests: await db
-          .select()
-          .from(ProfileFollowRequests)
-          .orderBy(asc(ProfileFollowRequests.id)),
-      });
-    const stateBefore = await readFollowGraphState();
-    const originalFetch = globalThis.fetch;
-    globalThis.fetch = (() => {
-      throw new Error('Remote follow graph reads must not fetch remote collections.');
-    }) as typeof fetch;
-
-    let result: GraphQLResult<FollowGraph>;
-
-    try {
-      result = await requestFollowGraph(
-        `read-only-remote-follow-graph@${remoteDomain}`,
-        auth.token,
-      );
-    } finally {
-      globalThis.fetch = originalFetch;
-    }
-
-    assertNoGraphQLErrors(result!);
-    assert.equal(result!.data?.profileByHandle?.followersCount, 61);
-    assert.equal(result!.data?.profileByHandle?.followingCount, 67);
-    assert.equal(await readFollowGraphState(), stateBefore);
   });
 
   test('reads active posts for profiles on another local instance', async () => {
@@ -1019,119 +749,6 @@ type GraphQLErrorResult = {
 type GraphQLResult<TData = Record<string, unknown>> = {
   data?: TData;
   errors?: GraphQLErrorResult[];
-};
-
-type FollowGraph = {
-  profileByHandle: {
-    followers: { edges: Array<{ node: { id: string } }> };
-    followersCount: number;
-    following: { edges: Array<{ node: { id: string } }> };
-    followingCount: number;
-    viewerFollow: { id: string } | null;
-    viewerState: { follow: { id: string } | null; isSelf: boolean } | null;
-  } | null;
-};
-
-const requestFollowGraph = (handle: string, token?: string) =>
-  requestGraphQL<FollowGraph>(
-    `query FollowGraph($handle: String!) {
-      profileByHandle(handle: $handle) {
-        followers(first: 10) { edges { node { id } } }
-        followersCount
-        following(first: 10) { edges { node { id } } }
-        followingCount
-        viewerFollow { id }
-        viewerState { isSelf follow { id } }
-      }
-    }`,
-    { handle },
-    token,
-  );
-
-const assertRemoteFollowGraphHidesCounterparts = async ({
-  followPolicy = ProfileFollowPolicy.OPEN,
-  instanceState,
-  profileState = ProfileState.ACTIVE,
-}: {
-  followPolicy?: ProfileFollowPolicy;
-  instanceState?: InstanceState;
-  profileState?: ProfileState;
-}) => {
-  const remoteInstance = await createRemoteInstance();
-  const remote = await createProfile({
-    handle: 'filtered-remote-follow-graph',
-    instanceId: remoteInstance.id,
-  });
-  const counterpartInstanceId = instanceState
-    ? await createRemoteInstance({
-        domain: 'filtered-counterpart.example',
-        state: instanceState,
-      }).then((instance) => instance.id)
-    : localInstanceId;
-  const follower = await createProfile({
-    followPolicy,
-    handle: 'filtered-follower',
-    instanceId: counterpartInstanceId,
-    state: profileState,
-  });
-  const followee = await createProfile({
-    followPolicy,
-    handle: 'filtered-followee',
-    instanceId: counterpartInstanceId,
-    state: profileState,
-  });
-  await db.insert(ProfileFollows).values([
-    { followerProfileId: follower.id, followeeProfileId: remote.id },
-    { followerProfileId: remote.id, followeeProfileId: followee.id },
-  ]);
-
-  const result = await requestFollowGraph(`filtered-remote-follow-graph@${remoteDomain}`);
-
-  assertNoGraphQLErrors(result);
-  assert.deepEqual(result.data?.profileByHandle, {
-    followers: { edges: [] },
-    followersCount: 0,
-    following: { edges: [] },
-    followingCount: 0,
-    viewerFollow: null,
-    viewerState: null,
-  });
-};
-
-const requestViewerFollowScenario = async (
-  relation: 'outgoing' | 'pending' | 'reverse',
-  { followersCount, followingCount }: { followersCount?: number; followingCount?: number } = {},
-) => {
-  const auth = await createAuthenticatedSession();
-  const remoteInstance = await createRemoteInstance();
-  const remote = await createProfile({
-    followersCount,
-    followingCount,
-    handle: `viewer-${relation}-remote`,
-    instanceId: remoteInstance.id,
-  });
-  let followId: string | null = null;
-
-  if (relation === 'outgoing') {
-    const follow = await db
-      .insert(ProfileFollows)
-      .values({ followerProfileId: auth.profile.id, followeeProfileId: remote.id })
-      .returning()
-      .then(firstOrThrow);
-    followId = follow.id;
-  } else if (relation === 'reverse') {
-    await db
-      .insert(ProfileFollows)
-      .values({ followerProfileId: remote.id, followeeProfileId: auth.profile.id });
-  } else {
-    await db.insert(ProfileFollowRequests).values([
-      { followerProfileId: auth.profile.id, followeeProfileId: remote.id },
-      { followerProfileId: remote.id, followeeProfileId: auth.profile.id },
-    ]);
-  }
-
-  const result = await requestFollowGraph(`viewer-${relation}-remote@${remoteDomain}`, auth.token);
-  return { followId, result };
 };
 
 const requestGraphQL = async <TData = Record<string, unknown>>(
